kernel: proc: Improve handling of 'nswitch'
- Extend 'nswitch' to 64-bits as it increments fast
- Ensure writes to 'nswitch' are atomic

+++ sys/include/sys/sched.h
@@ -42,7 +42,7 @@
  * @idle: Number of milliseconds idle
  */
 struct sched_cpu {
-    uint32_t nswitch;
+    uint64_t nswitch;
 };
 
 /*
